related files
Calculate related files when a file is added to the Working Set.
Related file calculation happens in the background.  This commit
includes changes to Async that allow an a serial task to happen in the
background.

The related files list shows up in the UI on the right-side of the
Working Set.  A loading throbber is added while the list is being
calculated.

The related files list is used to give related items priority in the
CSS inline editor list of matches.  There is a slight visual treatment
on matches that are related.

I removed the horizontal scroll bar from the Working Set.  Its an easy
thing to change, but it is a big improvement IMHO.

The structure change broke a unit test, so there is also an update to
keep that test passing.